refactor: move render pass builder to module
