+++ src/renderer/mod.rs
@@ -4,7 +4,8 @@ pub use drawable::*;
 mod tests;
 use crate::material::{PSOBindGroupResources, PSO};
 use crate::utils::ResourceManager;
-use crate::*;
+use dashi::utils::*;
+use crate::render_pass::*;
 use dashi::*;
 use sdl2::event::Event;
 use sdl2::keyboard::Keycode;
