+++ tests/sample_shadows/bin.rs
@@ -3,9 +3,10 @@ use dashi::utils::*;
 use dashi::*;
 use glam::*;
 use inline_spirv::include_spirv;
+use koji::render_pass::*;
 use koji::*;
-use sdl2::event::Event;
 use sdl2::keyboard::Keycode;
+use sdl2::event::Event;
 // Shader code is located in `shaders/` and included via `include_spirv!`.
 use std::f32::consts::PI;
 use std::time::Instant;
